pyproject.toml
[project]
name = "crypto-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=4.5.34",
    "gradio>=6.4.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.2",
]
